+++ src/main/java/rest/koios/client/backend/api/epoch/model/EpochInfo.java
@@ -18,7 +18,7 @@ public class EpochInfo {
     /**
      * Epoch number
      */
-    private Long epochNo = null;
+    private Integer epochNo = null;
 
     /**
      * Total output value across all transactions in epoch
